# server/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Loveraf API",
    "version": "1.0.0",
    "description": "Production API contract for the Loveraf Android app and web administration panel. Money values use integer minor units."
  },
  "servers": [
    { "url": "https://api.loveraf.com/api/v1", "description": "Production" },
    { "url": "http://127.0.0.1:3001/api/v1", "description": "Local development" }
  ],
...
  "paths": {
...
    "/admin/operations-summary": {
      "get": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Read actionable marketplace operations counts", "responses": { "200": { "description": "Operations summary", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OperationsSummary" } } } }, "403": { "$ref": "#/components/responses/Error" } } }
    },
...
    "/admin/content/{key}": {
      "post": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Update validated public content with optimistic concurrency and an audit reason", "parameters": [ { "$ref": "#/components/parameters/ContentKey" } ], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ContentUpdate" } } } }, "responses": { "201": { "description": "Content updated" }, "400": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" } } }
    },
...
    "/admin/finance-summary": {
      "get": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Read order payment and refund reconciliation totals", "responses": { "200": { "description": "Finance summary", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FinanceSummary" } } } }, "403": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/seller-finance": {
      "get": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "List seller payable balances and pending COD totals", "parameters": [ { "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Offset" }, { "name": "search", "in": "query", "schema": { "type": "string" } }, { "name": "status", "in": "query", "schema": { "type": "string" } } ], "responses": { "200": { "description": "Seller finance overview" }, "403": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/seller-finance/{sellerId}": {
      "get": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Read one seller's payable ledger, pending COD and payout history", "parameters": [ { "$ref": "#/components/parameters/SellerId" }, { "$ref": "#/components/parameters/Limit" } ], "responses": { "200": { "description": "Seller finance detail" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/coupons": {
      "get": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "List coupon campaigns", "responses": { "200": { "description": "Coupon campaigns" }, "403": { "$ref": "#/components/responses/Error" } } },
      "post": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Create an audited coupon campaign", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CouponInput" } } } }, "responses": { "201": { "description": "Coupon created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Coupon" } } } }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/coupons/{id}": {
      "patch": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Update an audited coupon campaign with optimistic concurrency", "parameters": [ { "$ref": "#/components/parameters/Id" } ], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CouponInput" } } } }, "responses": { "200": { "description": "Coupon updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Coupon" } } } }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/gift-cards": {
      "get": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "List masked gift cards", "responses": { "200": { "description": "Gift-card page" } } },
      "post": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Issue a duplicate-safe gift card and return its code once", "parameters": [ { "$ref": "#/components/parameters/IdempotencyKey" } ], "responses": { "201": { "description": "Gift card issued" }, "400": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/gift-cards/{id}/revoke": {
      "patch": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Revoke an unused gift card with an audit reason", "parameters": [ { "$ref": "#/components/parameters/Id" } ], "responses": { "200": { "description": "Gift card revoked" }, "409": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/memberships": {
      "get": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "List membership entitlements", "responses": { "200": { "description": "Membership page" } } },
      "post": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Grant a duplicate-safe audited controlled membership entitlement", "parameters": [ { "$ref": "#/components/parameters/IdempotencyKey" } ], "responses": { "201": { "description": "Membership granted" }, "400": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/memberships/{id}/revoke": {
      "patch": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Revoke an active membership with an audit reason", "parameters": [ { "$ref": "#/components/parameters/Id" } ], "responses": { "200": { "description": "Membership revoked" }, "409": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/affiliates": {
      "get": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Reconcile affiliate links, clicks, orders and commissions", "responses": { "200": { "description": "Affiliate report page" } } }
    },
    "/admin/affiliates/{id}": {
      "patch": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Enable or disable an affiliate link with an audit reason", "parameters": [ { "$ref": "#/components/parameters/Id" } ], "responses": { "200": { "description": "Affiliate link updated" }, "409": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/notification-campaigns": {
      "get": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "List notification campaigns", "responses": { "200": { "description": "Notification campaign page" } } },
      "post": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Create, schedule or immediately send an audited notification campaign", "responses": { "201": { "description": "Campaign created" } } }
    },
    "/admin/notification-campaigns/{id}": {
      "patch": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Update, send or cancel an undelivered notification campaign", "parameters": [ { "$ref": "#/components/parameters/Id" } ], "responses": { "200": { "description": "Campaign updated" }, "409": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/donation-campaigns": {
      "get": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "List donation campaigns", "responses": { "200": { "description": "Campaign page" } } },
      "post": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Create an audited donation campaign", "responses": { "201": { "description": "Campaign created" } } }
    },
    "/admin/donation-campaigns/{id}": {
      "patch": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Update a donation campaign with optimistic concurrency and an audit reason", "parameters": [ { "$ref": "#/components/parameters/Id" } ], "responses": { "200": { "description": "Campaign updated" }, "409": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/referrals": {
      "get": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Reconcile referral attribution, pending rewards and reversals", "parameters": [ { "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Offset" } ], "responses": { "200": { "description": "Referral reconciliation page" } } }
    },
    "/admin/loyalty": {
      "get": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Reconcile loyalty awards, redemptions and reversals", "parameters": [ { "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Offset" } ], "responses": { "200": { "description": "Loyalty reconciliation page" } } }
    },
    "/admin/rules": {
      "get": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "List versioned marketplace business rules", "responses": { "200": { "description": "Business rules" } } }
    },
    "/admin/rules/{key}": {
      "patch": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Update a validated business rule with optimistic concurrency and an audit reason", "parameters": [ { "$ref": "#/components/parameters/RuleKey" } ], "responses": { "200": { "description": "Business rule updated" }, "400": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/orders/{id}/status": {
      "patch": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Perform an audited operations order-status transition", "parameters": [ { "$ref": "#/components/parameters/Id" } ], "responses": { "200": { "description": "Order updated" }, "403": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/orders/{id}/shipment": {
      "patch": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Record audited manual courier tracking", "parameters": [ { "$ref": "#/components/parameters/Id" } ], "responses": { "200": { "description": "Shipment updated" }, "403": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/orders/{id}/cod-collection": {
      "post": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Record an idempotent reconciled COD collection", "parameters": [ { "$ref": "#/components/parameters/Id" }, { "$ref": "#/components/parameters/IdempotencyKey" } ], "responses": { "200": { "description": "COD collection recorded" }, "403": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/returns/{id}/status": {
      "patch": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Perform an audited after-sales transition with finance separation", "parameters": [ { "$ref": "#/components/parameters/Id" } ], "responses": { "200": { "description": "After-sales request updated" }, "403": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/reviews/{id}/status": {
      "patch": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Publish or reject a review with an audit reason", "parameters": [ { "$ref": "#/components/parameters/Id" } ], "responses": { "200": { "description": "Review moderated" }, "403": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/users/{id}/roles": {
      "patch": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Assign least-privilege administrator responsibilities and revoke existing sessions", "parameters": [ { "$ref": "#/components/parameters/Id" } ], "responses": { "200": { "description": "Administrator access updated" }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/{resource}": {
      "get": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "List one permission-scoped administration resource", "parameters": [ { "$ref": "#/components/parameters/AdminResource" }, { "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Offset" } ], "responses": { "200": { "description": "Administration resource page" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/{resource}/{id}": {
      "patch": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Moderate a supported resource with an audit reason", "parameters": [ { "$ref": "#/components/parameters/AdminModerationResource" }, { "$ref": "#/components/parameters/Id" } ], "responses": { "200": { "description": "Resource moderated" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/features": {
      "get": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "List super-admin-controlled capability switches", "responses": { "200": { "description": "Capability switches" }, "403": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/marketing-overview": {
      "get": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Read finance-scoped marketing readiness and metrics", "responses": { "200": { "description": "Marketing and rewards readiness overview" }, "403": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/features/{key}": {
      "patch": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Enable or disable a capability after validating prerequisites", "parameters": [ { "name": "key", "in": "path", "required": true, "schema": { "type": "string", "enum": ["rewards", "referral", "coupons", "affiliate", "loyalty", "gift_cards", "membership"] } } ], "responses": { "200": { "description": "Updated capability" }, "400": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" } } }
    },
...
    "/admin/withdrawals/{id}/status": {
      "patch": { "security": [ { "bearerAuth": [] } ], "tags": ["Administration"], "summary": "Approve, reject or record a completed payout", "parameters": [ { "$ref": "#/components/parameters/Id" } ], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WithdrawalStatusInput" } } } }, "responses": { "200": { "description": "Updated withdrawal", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Withdrawal" } } } }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" } } }
    }
  },
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } },
    "parameters": {
      "Id": { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^[a-fA-F0-9]{24}$" } },
      "Limit": { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 } },
      "Offset": { "name": "offset", "in": "query", "schema": { "type": "integer", "minimum": 0, "default": 0 } },
      "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "format": "uuid" } }
      ,"ContentKey": { "name": "key", "in": "path", "required": true, "schema": { "type": "string", "enum": ["categories", "home-banners", "home-promotion", "terms-policies", "about", "faq"] } }
...
      ,"SellerId": { "name": "sellerId", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^[a-fA-F0-9]{24}$" } }
...
      ,"RuleKey": { "name": "key", "in": "path", "required": true, "schema": { "type": "string", "enum": ["promo_usage", "referral", "withdrawal", "membership", "loyalty", "affiliate", "seller_commission"] } }
      ,"AdminResource": { "name": "resource", "in": "path", "required": true, "schema": { "type": "string", "enum": ["users", "sellers", "products", "orders", "returns", "reviews", "reports", "audit", "withdrawals"] } }
      ,"AdminModerationResource": { "name": "resource", "in": "path", "required": true, "schema": { "type": "string", "enum": ["users", "sellers", "products", "reports"] } }
    },
    "responses": { "Error": { "description": "Structured API error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } } },
    "schemas": {
...
      "ContentUpdate": { "type": "object", "required": ["data", "reason", "version"], "additionalProperties": false, "properties": { "data": { "oneOf": [ { "type": "array", "maxItems": 100 }, { "type": "object" } ] }, "reason": { "type": "string", "minLength": 3, "maxLength": 500 }, "version": { "type": "integer", "minimum": 0 } } },
      "Error": {
        "type": "object",
        "required": ["message", "error"],
        "properties": {
          "message": { "type": "string" },
          "fields": { "type": "object", "additionalProperties": true },
          "details": {},
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "fields": { "type": "object", "additionalProperties": true },
              "details": {}
            }
          }
        }
      },
...
      "OperationsSummary": { "type": "object", "required": ["id", "status", "pendingSellers", "pendingProducts", "stuckOrders", "uncollectedCod", "openReturns", "openReports", "generatedAt"], "properties": { "id": { "const": "operations-summary" }, "status": { "enum": ["healthy", "attention"] }, "pendingSellers": { "type": "integer", "minimum": 0 }, "pendingProducts": { "type": "integer", "minimum": 0 }, "stuckOrders": { "type": "integer", "minimum": 0 }, "uncollectedCod": { "type": "integer", "minimum": 0 }, "openReturns": { "type": "integer", "minimum": 0 }, "openReports": { "type": "integer", "minimum": 0 }, "generatedAt": { "type": "string", "format": "date-time" } } },
      "Coupon": { "type": "object", "required": ["id", "code", "name", "discountType", "discountValue", "minimumOrderMinor", "startsAt", "endAt", "totalLimit", "perUserLimit", "redemptionCount", "enabled", "version"], "properties": { "id": { "type": "string" }, "code": { "type": "string" }, "name": { "type": "string" }, "description": { "type": "string" }, "discountType": { "enum": ["percent", "fixed"] }, "discountValue": { "type": "integer" }, "minimumOrderMinor": { "type": "integer" }, "maximumDiscountMinor": { "type": "integer" }, "startsAt": { "type": "string", "format": "date-time" }, "endAt": { "type": "string", "format": "date-time" }, "totalLimit": { "type": "integer" }, "perUserLimit": { "type": "integer" }, "redemptionCount": { "type": "integer" }, "enabled": { "type": "boolean" }, "version": { "type": "integer" } } },
      "CouponInput": { "allOf": [ { "$ref": "#/components/schemas/Coupon" }, { "type": "object", "required": ["reason"], "properties": { "reason": { "type": "string", "minLength": 3 } } } ] },
...
      "FinanceSummary": { "type": "object", "required": ["id", "currency", "orderCount", "originalOrderMinor", "currentPayableMinor", "collectedMinor", "refundedMinor", "netCollectedMinor", "outstandingCodOrders", "promoLiabilityMinor", "earningsLiabilityMinor", "pendingEarningsMinor", "sellerPayableMinor", "openWithdrawalMinor", "openWithdrawalCount", "paidWithdrawalMinor", "generatedAt"], "properties": { "id": { "const": "finance-summary" }, "currency": { "const": "BDT" }, "orderCount": { "type": "integer", "minimum": 0 }, "originalOrderMinor": { "type": "integer" }, "currentPayableMinor": { "type": "integer" }, "collectedMinor": { "type": "integer" }, "refundedMinor": { "type": "integer" }, "netCollectedMinor": { "type": "integer" }, "outstandingCodOrders": { "type": "integer", "minimum": 0 }, "promoLiabilityMinor": { "type": "integer" }, "earningsLiabilityMinor": { "type": "integer" }, "pendingEarningsMinor": { "type": "integer" }, "sellerPayableMinor": { "type": "integer" }, "openWithdrawalMinor": { "type": "integer", "minimum": 0 }, "openWithdrawalCount": { "type": "integer", "minimum": 0 }, "paidWithdrawalMinor": { "type": "integer", "minimum": 0 }, "generatedAt": { "type": "string", "format": "date-time" } } },
...
      "WithdrawalStatusInput": { "type": "object", "required": ["status", "version", "reason"], "additionalProperties": false, "properties": { "status": { "enum": ["approved", "rejected", "paid"] }, "version": { "type": "integer", "minimum": 0 }, "reason": { "type": "string", "minLength": 3, "maxLength": 500 }, "settlement": { "type": "object", "required": ["reference", "paidAt"], "properties": { "reference": { "type": "string" }, "paidAt": { "type": "string", "format": "date-time" } } } } },
      "Withdrawal": { "type": "object", "required": ["id", "sourceKind", "amountMinor", "feeMinor", "payoutMinor", "currency", "destination", "status", "version"], "properties": { "id": { "type": "string" }, "sourceKind": { "enum": ["earnings", "seller_payable"] }, "sellerId": { "type": "string" }, "sellerName": { "type": "string" }, "amountMinor": { "type": "integer" }, "feeMinor": { "type": "integer" }, "payoutMinor": { "type": "integer" }, "currency": { "const": "BDT" }, "destination": { "type": "object" }, "status": { "enum": ["requested", "approved", "rejected", "paid"] }, "version": { "type": "integer" }, "settlement": { "type": "object" }, "createdAt": { "type": "string", "format": "date-time" } } }
    }
  }
}